requirements.txt
Django==3.2
Pillow==8.2.0
common-framework==2021.4.1
coreapi==2.3.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-compressor==2.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-htmlmin==0.11.0
flake8==3.9.0
folium==0.12.1
ipython==7.22.0
markdown==3.3.4
markdown2==2.4.0
python-dateutil==2.8.1
pytz==2021.1
